Set logging to DEBUG by default on tests
When tests fail in travis-ci the logging is pretty sparse, it defaults
to INFO by default, and there are no timestamps in the logs. This does
some very basic logging config to add that information into tests. The
alternative would be to load pytest fixtures, but that would require
touching lots of tests directly, where this provides a global
solution.